# swag_docs/swagger.yaml
definitions:
...
  accounts.ExistingCheck:
    properties:
      cf_token:
        type: string
      email:
        type: string
    type: object
  accounts.ExistingResponse:
    properties:
      exists:
        type: boolean
    type: object
...
  accounts.SignupResponse:
    properties:
      redirect_url:
        type: string
      token:
        type: string
      verification_token:
        type: string
    type: object
...
  accounts.VerifyInput:
    properties:
      email:
        type: string
      verify:
        type: string
    type: object
  accounts.VerifyResponse:
    properties:
      redirect_url:
        type: string
      token:
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      error:
        type: string
      success:
        type: boolean
    type: object
  response.SuccessResponse:
    properties:
      data: {}
      success:
        type: boolean
    type: object
info:
  contact: {}
paths:
...
  /account/check:
    post:
      consumes:
      - application/json
      description: Checks whether an email address is already registered
      parameters:
      - description: Email to check
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/accounts.ExistingCheck'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/accounts.ExistingResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Check existing email
      tags:
      - Account
...
  /signup:
    post:
      consumes:
      - application/json
      description: Creates a new account with email and password or completes invited
        user signup
      parameters:
      - description: Signup data
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/accounts.SignupResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Standard signup
      tags:
      - Account
  /signup/oauth:
    post:
      consumes:
      - application/json
      description: Completes signup process for OAuth authenticated users
      parameters:
      - description: Signup data with OAuth token
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/accounts.SignupResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: OAuth signup
      tags:
      - Account
  /verify/email:
    post:
      consumes:
      - application/json
      description: Verifies a user's email address using the verification key and
        creates a user session
      parameters:
      - description: Verification details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/accounts.VerifyInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/accounts.VerifyResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Verify email address
      tags:
      - Account
  /verify/invite:
    post:
      consumes:
      - application/json
      description: Verifies an account invitation using the verification key and creates
        a user session
      parameters:
      - description: Verification details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/accounts.VerifyInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/accounts.VerifyResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Verify account invite
      tags:
      - Account
swagger: "2.0"